requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
Fastapi
uvicorn
groq
openai